Restructure failover test
