+++ infra/gravity/cluster_status.go
@@ -193,21 +193,3 @@ func apiserverNode(ctx context.Context, nodes []Gravity) (api Gravity, other []G
 
 	return api, nodes[1:], nil
 }
-
-// GetLeaderNode finds and returns the leader node.
-func (c *TestContext) GetLeaderNode(nodes []Gravity) (leader Gravity, err error) {
-	ctx, cancel := context.WithTimeout(c.ctx, c.timeouts.Status)
-	defer cancel()
-	for _, node := range nodes {
-		if node.IsLeader(ctx) {
-			if leader != nil {
-				return nil, trace.BadParameter("multiple leader nodes")
-			}
-			leader = node
-		}
-	}
-	if leader == nil {
-		return nil, trace.NotFound("unable to get leader node")
-	}
-	return leader, nil
-}
